[FIX] attribute_set: fix create options w/ rel model
When you define an attribute as a select field and add a related model,
you have the option to "Load attribute options".

On this wizard, an "option_ids" dummy field is created via
fields_view_get.

This change deletes option_ids from create (once the options are created) and read